requirements.txt
pandas
openpyxl
pyperclip
Pillow

#Digite no seu terminal: pip install -r requirements.txt
#Isso instalará todas as bibliotecas essenciais para o funcionamento do nosso sistema